// admin/src/models/TaxiReservationResponse.ts
export interface TaxiReservationResponse {
  id: string
  userId: string
  userName: string
  departurePlace: string
  arrivalPlace: string
  userPhoneNumber: string
  userNumberOfPassenger: number
  userPassengers: string
  numberOfTickets: number
  reservationStatus: number
  reservationDatetime: string
  latestUpdateDatetime: string
}